+++ linopy/model.py
@@ -28,12 +28,12 @@ from xarray.core.types import T_Chunks
 
 from linopy import solvers
 from linopy.common import (
-    align_to_coords,
     as_dataarray,
     assign_multiindex_safe,
     best_int,
     maybe_replace_signs,
     replace_by_map,
+    strict_broadcast_to_coords,
     to_path,
 )
 from linopy.constants import (
@@ -774,8 +774,12 @@ class Model:
                     "Semi-continuous variables require a positive scalar lower bound."
                 )
 
-        lower_da = align_to_coords(lower, coords, label="lower bound", **kwargs)
-        upper_da = align_to_coords(upper, coords, label="upper bound", **kwargs)
+        lower_da = strict_broadcast_to_coords(
+            lower, coords, label="lower bound", **kwargs
+        )
+        upper_da = strict_broadcast_to_coords(
+            upper, coords, label="upper bound", **kwargs
+        )
         data = Dataset(
             {
                 "lower": lower_da,
@@ -788,7 +792,7 @@ class Model:
         self._check_valid_dim_names(data)
 
         if mask is not None:
-            mask = align_to_coords(
+            mask = strict_broadcast_to_coords(
                 mask,
                 coords if coords is not None else data.coords,
                 label="mask",
@@ -1057,7 +1061,9 @@ class Model:
         (data,) = xr.broadcast(data, exclude=[TERM_DIM])
 
         if mask is not None:
-            mask = align_to_coords(mask, data.coords, label="mask").astype(bool)
+            mask = strict_broadcast_to_coords(mask, data.coords, label="mask").astype(
+                bool
+            )
 
         # Auto-mask based on null expressions or NaN RHS (use numpy for speed)
         if self.auto_mask: